Refactor: 스타일드 컴포넌트 파일 분리(#29)
스타일드 컴포넌트 파일 분리 및 파일 명 변경
