+++ src/components/common/ProductItem/ProductItem.styled.js
@@ -1,7 +1,6 @@
-import React from 'react';
 import { styled } from 'styled-components';
 
-const ItemBox = styled.article`
+export const ItemBox = styled.article`
   width: 14rem;
   height: 13.2rem;
 
@@ -27,15 +26,3 @@ const ItemBox = styled.article`
     margin-top: 0.4rem;
   }
 `;
-
-  export default function ProductItem({itemName,price,itemImg}) {
-  return (
-    <ItemBox>
-
-      <img src={itemImg} alt="상품 사진" />
-      <h1>{itemName}</h1>
-      <p>{price}</p>
-    </ItemBox>
-  );
-}
-
