+++ src/pages/NotFound/NotFound.jsx
@@ -1,22 +1,7 @@
 import React from 'react';
-import styled from 'styled-components';
 import errorImg from '../../assets/img/404.svg';
 import StyledBtn from '../../components/common/Button/Button';
-
-const NotFoundPageWrap = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  margin-top: 27.4rem;
-  gap: 2rem;
-  img {
-    width: 19.6rem;
-  }
-  p {
-    font-size: 1.4rem;
-  }
-`;
+import NotFoundPageWrap from './NotFound.styled';
 
 export default function NotFound() {
   return (
