+++ src/pages/NotFound/NotFound.styled.js
@@ -1,0 +1,17 @@
+import styled from 'styled-components';
+
+const NotFoundPageWrap = styled.div`
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  margin-top: 27.4rem;
+  gap: 2rem;
+  img {
+    width: 19.6rem;
+  }
+  p {
+    font-size: 1.4rem;
+  }
+`;
+export default NotFoundPageWrap;
